feat: allow decimal values (1 digit behind comma) for all odometer readings and inputs

+++ src/app/api/fuel-logs/[id]/route.ts
@@ -63,7 +63,7 @@ export async function PUT(
       );
     }
 
-    const parsedOdo = parseInt(odometer);
+    const parsedOdo = parseFloat(odometer);
     const parsedLiters = parseFloat(liters);
     const parsedPrice = parseInt(price);
